# planner: use `BatchPointGet` to improve the `SELECT ...WHERE IN` performance