--- a/planner/core/point_get_plan_test.go
+++ b/planner/core/point_get_plan_test.go
@@ -221,12 +221,21 @@ func (s *testPointGetSuite) TestWhereIn2BatchPointGet(c *C) {
 		"└─Point_Get_2 1.00 root table:t, index:a b c",
 	))
 
-	tk.MustQuery("explain select * from t where a in (1, 2, 3, 1, 2)").Check(testkit.Rows(
+	tk.MustQuery("explain select * from t where a in (1, 2, 3, 4, 5)").Check(testkit.Rows(
 		"Union_6 5.00 root ",
 		"├─Point_Get_1 1.00 root table:t, handle:1",
 		"├─Point_Get_2 1.00 root table:t, handle:2",
 		"├─Point_Get_3 1.00 root table:t, handle:3",
 		"├─Point_Get_4 1.00 root table:t, handle:4",
 		"└─Point_Get_5 1.00 root table:t, handle:5",
+	))
+
+	tk.MustQuery("explain select * from t where a in (1, 2, 3, 1, 2)").Check(testkit.Rows(
+		"Union_6 5.00 root ",
+		"├─Point_Get_1 1.00 root table:t, handle:1",
+		"├─Point_Get_2 1.00 root table:t, handle:2",
+		"├─Point_Get_3 1.00 root table:t, handle:3",
+		"├─Point_Get_4 1.00 root table:t, handle:1",
+		"└─Point_Get_5 1.00 root table:t, handle:2",
 	))
 }