@lonng 2019-08-20
/run-all-tests